Small cleanup and add yaml config option

+++ shared/configs/game_config.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 @dataclass
 class GameConfig:
     title: str
-    height: int
-    width: int
+    height: float
+    width: float
